% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_holidays.R
\name{swiss_holidays}
\alias{swiss_holidays}
\title{Swiss holidays (Ferien und Feiertage)}
\usage{
swiss_holidays(
  canton = ivttools:::canton,
  year,
  holidays = TRUE,
  feiertage = TRUE
)
}
\arguments{
\item{canton}{character: one of ivttools::canton}

\item{year}{character}

\item{holidays}{logical}

\item{feiertage}{logical}
}
\value{
list of dfs
}
\description{
Swiss holidays (Ferien und Feiertage)
}
