% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_weather.R
\name{swiss_weather_codebook}
\alias{swiss_weather_codebook}
\title{Codebook for swiss weather data}
\usage{
swiss_weather_codebook()
}
\description{
Codebook for swiss weather data
}
\seealso{
\code{swiss_weather_data()}
}
