% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_weather_data.R
\name{swiss_weather_data}
\alias{swiss_weather_data}
\title{Download Swiss weather data}
\usage{
swiss_weather_data()
}
\description{
Download Swiss weather data
}
\seealso{
\code{swiss_weather_codebook()}
}
